+++ src/helpers/prendyHelpers/helpers.ts
@@ -38,65 +38,4 @@ import { hideSticker, moveSticker, showSticker } from "./stickers";
 
 // function doThis
 
-export function makePrendyStoryHelpers<T_MyTypes extends MyTypes = MyTypes>() {
-  const prendyAssets = meta.assets!;
-  const prendyStores = meta.stores!;
-  const storeHelpers = meta.repond!;
-
-  return {
-    characters: {
-      setCharAnimation,
-      setCharPosition,
-      setCharRotationY,
-      springCharRotation,
-      springAddToCharRotationY,
-      lookAtOtherCharacter,
-      lookAtEachother,
-      moveCharacterAt2DAngle,
-    },
-    dolls: {
-      setDollPosition,
-      setDollRotation,
-      setDollRotationY,
-      springDollRotationY,
-      springAddToDollRotationY,
-      pushDollRotationY,
-      lookAtOtherDoll,
-      setDollAnimation,
-      focusOnDoll,
-      setDollToSpot,
-      springDollToSpot,
-      dollLooksAtSpot,
-      moveDollAt2DAngle,
-      hideDoll,
-      toggleDollMeshes,
-      getDollBonePosition,
-      // springDollPosition,
-      // slideDollPosition,
-    },
-    players: {
-      enableMovement,
-      isHolding,
-      takePickup,
-      setPlayerAnimations,
-    },
-    scene: {
-      // changeSegmentAtLoop,
-      // changeCameraAtLoop,
-      lookAtSpot,
-      hideWallIf,
-      showStoryView,
-      setSegment,
-      setCamera,
-      goToNewPlace,
-    },
-    sound: { playNewMusic, stopAllMusic, playSound, stopSound, stopAllSounds },
-    speech: {
-      showSpeech,
-      showMiniBubble,
-      hideMiniBubble,
-      showAlarmText,
-    },
-    stickers: { moveSticker, showSticker, hideSticker },
-  };
-}
+export function makePrendyStoryHelpers<T_MyTypes extends MyTypes = MyTypes>() {}
